Migration to RX java 2
- removed rx 1.2.6 dependency
- removed emmiter, not needed. I use direcctly the Observable emmiter, without unneded absttractions
