+++ library/build.gradle
@@ -14,7 +14,7 @@ android {
 
 dependencies {
   compile "com.android.support:support-annotations:$supportLibraryVersion"
-  compile 'io.reactivex:rxjava:1.2.5'
+  compile 'io.reactivex.rxjava2:rxjava:2.0.1'
 
   testCompile 'junit:junit:4.12'
   testCompile 'org.mockito:mockito-core:1.10.19'
