+++ library/src/test/java/com/jenzz/appstate/RxAppStateMonitorTest.java
@@ -7,7 +7,8 @@ import com.jenzz.appstate.stubs.StubAppStateRecognizer;
 
 import org.junit.Test;
 
-import rx.observers.TestSubscriber;
+
+import io.reactivex.observers.TestObserver;
 
 import static com.jenzz.appstate.AppState.BACKGROUND;
 import static com.jenzz.appstate.AppState.FOREGROUND;
@@ -61,13 +62,13 @@ public class RxAppStateMonitorTest {
   @Test
   public void emitsAppStates() {
     FakeApplication fakeApplication = new FakeApplication();
-    TestSubscriber<AppState> subscriber = TestSubscriber.create();
+    TestObserver<AppState> subscriber = TestObserver.create();
     RxAppStateMonitor.monitor(fakeApplication).subscribe(subscriber);
 
     fakeApplication.goForeground();
     fakeApplication.goBackground();
 
     subscriber.assertValues(FOREGROUND, BACKGROUND);
-    subscriber.assertNoTerminalEvent();
+    subscriber.assertNotComplete();
   }
 }
