+++ library/src/test/java/com/jenzz/appstate/stubs/StubAppStateRecognizer.java
@@ -9,24 +9,11 @@ import com.jenzz.appstate.internal.AppStateRecognizer;
 import java.util.ArrayList;
 import java.util.List;
 
-import rx.functions.Action1;
 
 import static com.jenzz.appstate.AppState.BACKGROUND;
 
 public class StubAppStateRecognizer implements AppStateRecognizer {
 
-  public static final Action1<AppStateListener> ACTION_FOREGROUND = new Action1<AppStateListener>() {
-    @Override
-    public void call(AppStateListener appStateListener) {
-      appStateListener.onAppDidEnterForeground();
-    }
-  };
-  public static final Action1<AppStateListener> ACTION_BACKGROUND = new Action1<AppStateListener>() {
-    @Override
-    public void call(AppStateListener appStateListener) {
-      appStateListener.onAppDidEnterBackground();
-    }
-  };
 
   @NonNull private final List<AppStateListener> listeners = new ArrayList<>();
 
@@ -72,9 +59,4 @@ public class StubAppStateRecognizer implements AppStateRecognizer {
     return isStarted;
   }
 
-  public void notifyAppStateListener(@NonNull Action1<AppStateListener> action) {
-    for (AppStateListener listener : listeners) {
-      action.call(listener);
-    }
-  }
 }
